# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/import/trades": {
      "post": {
        "operationId": "ImportController_importTrades",
        "summary": "Import / Re-import trades CSV",
        "description": "Upload a multipart trades.csv file. Validates the entire file (columns, data types, ranges, duplicates, short positions) before atomic replacement in the database.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "CSV file with trades data (trade_id, timestamp, exchange, symbol, side, quantity, price_usd, fee_usd)",
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/FileUploadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Trades imported successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResultDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed. No partial state saved (atomic rejection).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "CSV Import & Management"
        ]
      }
    },
    "/api/import/prices": {
      "post": {
        "operationId": "ImportController_importPrices",
        "summary": "Import / Re-import prices CSV",
        "description": "Upload a multipart prices.csv file (as_of, symbol, price_usd). Atomically updates stored price snapshot.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "CSV file with price snapshot data (as_of, symbol, price_usd)",
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/FileUploadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Prices imported successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResultDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed. No partial state saved.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "CSV Import & Management"
        ]
      }
    },
...
  },
  "info": {
    "title": "Crypto Portfolio Analytics API",
    "description": "Interactive OpenAPI / Swagger specification for Crypto Portfolio Analytics, Holdings, and Transaction Explorer.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "FileUploadDto": {
        "type": "object",
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "description": "CSV file to upload"
          }
        },
        "required": [
          "file"
        ]
      },
      "ImportResultDto": {
        "type": "object",
        "properties": {
          "imported": {
            "type": "number",
            "description": "Number of rows successfully imported and saved",
            "example": 200
          }
        },
        "required": [
          "imported"
        ]
      },
      "ValidationErrorDto": {
        "type": "object",
        "properties": {
          "row": {
            "type": "number",
            "description": "Row index in CSV file (1-based, or 0 for header, -1 for cross-row)",
            "example": 5
          },
          "field": {
            "type": "string",
            "description": "Name of the invalid column/field",
            "example": "quantity"
          },
          "message": {
            "type": "string",
            "description": "Actionable error description",
            "example": "quantity must be greater than zero"
          }
        },
        "required": [
          "row",
          "message"
        ]
      },
      "ImportErrorResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error summary message",
            "example": "Import failed validation"
          },
          "errors": {
            "description": "List of specific validation errors found in the CSV",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationErrorDto"
            }
          }
        },
        "required": [
          "message",
          "errors"
        ]
      },
...
    }
  }
}